+++ src/engine/Engine.cpp
@@ -160,7 +160,7 @@ GL_PREFERRED_TEXTURE_FORMAT_RGB8:  0x{:x}
             elapsedDuration = loopTime - prevLoopTime;
 
             clock.frameCount += 1;
-            clock.ts = static_cast<float>(glfwGetTime());
+            clock.ts = static_cast<double>(glfwGetTime());
             clock.elapsedSecs = elapsedDuration.count();
 
             if (m_registry->m_pendingSnapshotRegistry->isDirty()) {
